+++ app/serializers/album_photos.js
